requirements.txt
appdirs==1.4.3
appnope==0.1.2
asgiref==3.2.5
astroid==2.4.2
attrs==19.3.0
backcall==0.2.0
black==20.8b1
certifi==2020.12.5
click==7.1.2
decorator==4.4.2
distlib==0.3.0
Django==3.0.4
django-cors-headers==3.5.0
django-crispy-forms==1.9.1
django-phone-field==1.8.1
djangorestframework==3.11.1
elasticsearch==7.10.1
elasticsearch-dsl==7.3.0
filelock==3.0.12
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.5.2
jedi==0.17.2
jupyter-client==6.1.7
jupyter-core==4.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.3
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2019.3
pyzmq==20.0.0
regex==2020.7.14
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.3
virtualenv==20.0.9
wcwidth==0.2.5
wrapt==1.12.1
